Add exercises 4.18 - 4.21

+++ controllers/blogs.js
@@ -1,27 +1,16 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
-
-blogsRouter.get('/', async (request, response, next) => {
-  try {
-    const blogs = await Blog
-      .find({})
-      .populate('user', { username: 1, name: 1 })
-    response.json(blogs)
-  }
-  catch (exception) {
-    next(exception)
-  }
-})
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/:id', async (request, response, next) => {
-  try{
+  try {
     const blog = await Blog
       .findById(request.params.id)
       .populate('user', { username: 1, name: 1 })
 
     if (blog) {
-      response.json(blog.toJSON())
+      return response.json(blog.toJSON())
     } else {
       response.status(404).end()
     }
@@ -30,17 +19,47 @@ blogsRouter.get('/:id', async (request, response, next) => {
   }
 })
 
-blogsRouter.post('/', async (request, response, next) => {
+blogsRouter.get('/', async (request, response, next) => {
   try {
-    const user = await User.findById(request.body.user)
+    const blogs = await Blog
+      .find({})
+      .populate('user', { username: 1, name: 1 })
+    return response.json(blogs)
+  }
+  catch (exception) {
+    next(exception)
+  }
+})
 
-    const blog = new Blog(request.body)
-    const result = await blog.save()
+blogsRouter.post('/', async (request, response, next) => {
+  const body = request.body
+  const token = request.token
 
-    user.blogs = user.blogs.concat(blog._id)
+  if (!token) {
+    return response.status(401).end()
+  }
+
+  try {
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+
+    const blog = new Blog({
+      title: body.title,
+      author: body.author || user.name,
+      url: body.url,
+      likes: body.likes,
+      user: user._id,
+    })
+
+    const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
     await user.save()
-
-    response.status(201).json(result.toJSON())
+    response.status(201).json(savedBlog.toJSON())
   }
   catch (exception) {
     next(exception)
@@ -60,7 +79,32 @@ blogsRouter.put('/:id', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const token = request.token
+
+  if (!token) {
+    return response.status(401).end()
+  }
+
   try {
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+    const userJson = user.toJSON()
+
+    const blog = await Blog
+      .findById(request.params.id)
+      .populate('user', { username: 1, name: 1 })
+
+    const blogJson = blog.toJSON()
+
+    if (blogJson.user.id !== userJson.id) {
+      return response.status(403).json({ error: 'Cannot delete blog posts of other users' })
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
   } catch (exception) {
